Move MultiLabelMetrics to metrics
